# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Security Scanner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type CreateScanBodyOptions = {
  checkHeaders: boolean;
  checkCors: boolean;
  checkSqlInjection: boolean;
  checkXss: boolean;
  checkOpenRedirect: boolean;
  checkHttpMethods: boolean;
  checkRateLimit: boolean;
};

export interface CreateScanBody {
  /** Target URL to scan */
  url: string;
  options: CreateScanBodyOptions;
}

export type ScanStatusProperty =
  (typeof ScanStatusProperty)[keyof typeof ScanStatusProperty];

export const ScanStatusProperty = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export type ScanRiskLevel =
  | (typeof ScanRiskLevel)[keyof typeof ScanRiskLevel]
  | null;

export const ScanRiskLevel = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
  info: "info",
} as const;

export interface Scan {
  id: string;
  url: string;
  status: ScanStatusProperty;
  riskScore: number | null;
  riskLevel: ScanRiskLevel;
  vulnerabilityCount: number;
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
  createdAt: string;
  completedAt: string | null;
}

export interface ScanList {
  scans: Scan[];
  total: number;
}

export type VulnerabilityType =
  (typeof VulnerabilityType)[keyof typeof VulnerabilityType];

export const VulnerabilityType = {
  sql_injection: "sql_injection",
  xss: "xss",
  open_redirect: "open_redirect",
  cors_misconfiguration: "cors_misconfiguration",
  missing_security_header: "missing_security_header",
  http_method_abuse: "http_method_abuse",
  command_injection: "command_injection",
  idor: "idor",
  rate_limit_bypass: "rate_limit_bypass",
  information_disclosure: "information_disclosure",
} as const;

export type VulnerabilitySeverity =
  (typeof VulnerabilitySeverity)[keyof typeof VulnerabilitySeverity];

export const VulnerabilitySeverity = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
  info: "info",
} as const;

export interface Vulnerability {
  id: string;
  scanId: string;
  type: VulnerabilityType;
  severity: VulnerabilitySeverity;
  title: string;
  description: string;
  affectedEndpoint: string;
  payload: string | null;
  evidence: string | null;
  aiExplanation: string | null;
  remediation: string;
  cvssScore: number | null;
  discoveredAt: string;
}

export type ScanEventEventType =
  (typeof ScanEventEventType)[keyof typeof ScanEventEventType];

export const ScanEventEventType = {
  request: "request",
  response: "response",
  finding: "finding",
  info: "info",
  error: "error",
} as const;

export type ScanEventSeverity =
  | (typeof ScanEventSeverity)[keyof typeof ScanEventSeverity]
  | null;

export const ScanEventSeverity = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
  info: "info",
} as const;

export interface ScanEvent {
  id: string;
  scanId: string;
  stage: string;
  eventType: ScanEventEventType;
  title: string;
  description: string;
  payload: string | null;
  requestHeaders: string | null;
  responseStatus: number | null;
  responseHeaders: string | null;
  responseBody: string | null;
  severity: ScanEventSeverity;
  timestamp: string;
}

export interface ScanDetail {
  scan: Scan;
  vulnerabilities: Vulnerability[];
  events: ScanEvent[];
}

export type ScanStatusStatus =
  (typeof ScanStatusStatus)[keyof typeof ScanStatusStatus];

export const ScanStatusStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface ScanStatus {
  id: string;
  status: ScanStatusStatus;
  /** Progress percentage 0-100 */
  progress: number;
  currentStage: string | null;
  stagesCompleted: string[];
  stagesTotal: number;
}

export interface ScanEventList {
  events: ScanEvent[];
}

export interface VulnerabilityList {
  vulnerabilities: Vulnerability[];
}

export type StatsSummaryTopVulnerabilityTypesItem = {
  type: string;
  count: number;
};

export interface StatsSummary {
  totalScans: number;
  totalVulnerabilities: number;
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
  avgRiskScore: number;
  topVulnerabilityTypes: StatsSummaryTopVulnerabilityTypesItem[];
}

export interface RecentActivity {
  recentScans: Scan[];
  recentVulnerabilities: Vulnerability[];
}

export type ListScansParams = {
  limit?: number;
  offset?: number;
};