# lib/api-zod/src/generated/types/createScanBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Security Scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateScanBodyOptions } from "./createScanBodyOptions";

export interface CreateScanBody {
  /** Target URL to scan */
  url: string;
  options: CreateScanBodyOptions;
}