# lib/api-zod/src/generated/types/scan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Security Scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanRiskLevel } from "./scanRiskLevel";
import type { ScanStatusProperty } from "./scanStatusProperty";

export interface Scan {
  id: string;
  url: string;
  status: ScanStatusProperty;
  riskScore: number | null;
  riskLevel: ScanRiskLevel;
  vulnerabilityCount: number;
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
  createdAt: string;
  completedAt: string | null;
}