# lib/api-zod/src/generated/types/scanDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Security Scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { Scan } from "./scan";
import type { ScanEvent } from "./scanEvent";
import type { Vulnerability } from "./vulnerability";

export interface ScanDetail {
  scan: Scan;
  vulnerabilities: Vulnerability[];
  events: ScanEvent[];
}