# lib/api-zod/src/generated/types/scanEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Security Scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanEventEventType } from "./scanEventEventType";
import type { ScanEventSeverity } from "./scanEventSeverity";

export interface ScanEvent {
  id: string;
  scanId: string;
  stage: string;
  eventType: ScanEventEventType;
  title: string;
  description: string;
  payload: string | null;
  requestHeaders: string | null;
  responseStatus: number | null;
  responseHeaders: string | null;
  responseBody: string | null;
  severity: ScanEventSeverity;
  timestamp: string;
}