# lib/api-zod/src/generated/types/scanEventList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Security Scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanEvent } from "./scanEvent";

export interface ScanEventList {
  events: ScanEvent[];
}