# lib/api-zod/src/generated/types/scanStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Security Scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanStatusStatus } from "./scanStatusStatus";

export interface ScanStatus {
  id: string;
  status: ScanStatusStatus;
  /** Progress percentage 0-100 */
  progress: number;
  currentStage: string | null;
  stagesCompleted: string[];
  stagesTotal: number;
}