# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Security Scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { StatsSummaryTopVulnerabilityTypesItem } from "./statsSummaryTopVulnerabilityTypesItem";

export interface StatsSummary {
  totalScans: number;
  totalVulnerabilities: number;
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
  avgRiskScore: number;
  topVulnerabilityTypes: StatsSummaryTopVulnerabilityTypesItem[];
}